Pixmap loader adds images to unreferenced list when cache: false
Don't add uncached images to the unreferenced list, since
they cannot be reused later as they are not in the cache.

Additionally, we currently search for images in the cache, even if we
set cache: false. Setting cache false should not put images in the cache
and should not use images that are in the cache.

Task-number: QTBUG-26676
Change-Id: Ib7eb42199ae6ae6154b696c83ad1dd959e0f208f
Reviewed-by: Andrew den Exter <andrew.den-exter@nokia.com>
Reviewed-by: Martin Jones <martin.jones@nokia.com>