Restore QML/C++ extensions tutorial and fix plugin docs and examples
This restores the "Writing QML Extensions with C++" tutorial that
was removed during the recent QML doc restructure.

This also updates outdated docs in "Creating C++ Plugins for QML"
that referred to the use of Q_EXPORT_PLUGIN2() and such as the
process for writing QML plugins in Qt 5 has changed.

Note the plugins.json line was removed from
examples/qml/cppextensions/plugins/plugins.pro since this json file
was removed in a previous commit, as plugins no longer require the
empty json file.

Change-Id: I795d6faf6a741466a952feb8e15b028ec3d52b69
Reviewed-by: Chris Adams <christopher.adams@nokia.com>