Clean up structure of JavaScript related docs
The global object and list of javascript functions should be linked to
as subsections from the "JavaScript Host Environment" topic page, and
all of the subsections should be visible from the module index page.

Also move the "restrictions" docs to this topic page as it's more
relevant to this section.

Also includes some minor doc improvements.

Task-number: QTBUG-26380

Change-Id: Iba6e4e510191bf5e4603d6cbb2826ad7bcdbd10c
Reviewed-by: Chris Adams <christopher.adams@nokia.com>