Use the Module Identifier Directive to make the TimeExample correct
From the documentation:

"Declares the module identifier of the module. The <ModuleIdentifier> is the
(dotted URI notation) identifier for the module, which must match the module's
install path.

The module identifier directive must be the first line of the file. Exactly one
module identifier directive may exist in the qmldir file."

More information can be found here: Task-number: QTBUG-26551

Change-Id: Ifd2a7e9a2868ac05baed0dbde9638ce85ff5004a
Reviewed-by: Matthew Vogt <mattvogt2@gmail.com>