Port to ia32 (stack based value passing)
In the architecture section, introduce a VALUE_FITS_IN_REGISTER
macro as well as a RegisterSize constant.

When storing the return value into a IR::Temp, either copy straight
from the return value register if the architecture supports values
in registers, otherwise do a load/store copy.

Added push() overload that allows pushing constant values.

push() with an IR::Temp* now pushes the actual value onto the stack,
because our run-time functions expect values to be passed by... value :)

Extended the function call logic to reserve space on the stack for
the function return value if it doesn't fit into a register. In that
case at least the ia32 ABI specifies that there is a hidden first
argument which is the address of the space where the callee can place
the return value itself.

Change-Id: Ice9435dfc1482f4045a4ec4ad2c9ab58d4dfc7aa
Reviewed-by: Lars Knoll <lars.knoll@digia.com>