Accessibility: Improve interface handling in Quick
This changes how interfaces are handled:
the QAccessibleQuickItem will simply subclass
all interfaces and dynamically return 0 or the
appropriate interface_cast.

This makes our implementation a lot more flexible.
To make use of the value interface, only a value
property is needed (together with a corresponding role).

Since the implementation of the interfaces became simpler,
the value interface and some text interface improvements
go along with the change.

Change-Id: I003ec3016d48d730a4acac467bce322167842f4d
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>