qmlscene can start with either an Item or Window as the qml root
Before, it assumed that the root is an Item and needs to have a
Window created.  But it's useful for an application to have a Window
as the root, and it was already possible by writing a different C++
main function (see qtdeclarative/examples/window/window/window.cpp).
It doesn't take much to give qmlscene this flexibility too.

Change-Id: Ie808e78a42074e13aa9d3c87723ec9ac8fdbaf4a
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>