Fix function calls on linux/ia32
Remove supposed stack size alignment correction after function calls

This code was supposed to "align" the stack pointer after a function call,
but we can't really add more to the stack pointer than what we previously
subtracted.

It should've practically been a noop on amd64 since we don't actually
use the stack for function calls there, which is why this manifested only
on ia32.

Change-Id: I5644e9e179bd1b7c45daca7d88687bd5d83a4104
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>