Enable the lexer to run in JS strict/non-strict mode as well as QML mode
If the lexer is running in JS mode (non-qmlMode), then allow for
FutureReservedWords as identifiers. Proper strict mode handling of them
should be implemented in the semantic phase.

QML mode is not affected, so it will reject most FutureReservedWords,
and a number of other keywords.

See also ECMA-262, version 5.1, paragraph 7.6.1.2.

Change-Id: Ide73a041f2dfda2de3b719d6437d517f469eb450
Reviewed-by: Lars Knoll <lars.knoll@digia.com>