Only free context if the owning QV8ContextResource gets destroyed
Since fdeee38b781376012c4f086276c3c376726c8839 QQmlXMLHttpRequest stores
the calling context for later use. This leads to issues after the first
request completes and the wrapping QV8ContextResource gets freed by
garbage collection and therefore removes the associated QQmlDataContext
which may still be required for later calls (e.g. if the calling context
is part of a stateless library).

This patch introduces an ownership flag for QV8ContextResource which
indicates if the associated context should be cleared when the object is
destroyed.

Task-number: QTBUG-28351

Change-Id: I552ebb5c55b889eb33f3884283c8fdf037ac33be
Reviewed-by: Alan Alpert <aalpert@rim.com>