Accessibility: Work with QQuickWindow
Before we would only be accessible when QQView was used.
While the window has the content item, that one will not
appear in the hierarchy of accessible objects since it has
no valueable properties to expose. Instead it's first child
will be used (which was the case before when QQView was used).

Change-Id: I03cefb5c8dcd8296d51ff2bfcb677386a5f1d698
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>