Fix quadratic behavior in the memory manager
The old freeList implementation was causing quadratic
behavior in alloc(), as the free item ended up in the
highest chunk.

The new implementation uses a fixed size array for small objects
(up to 256 bytes), a QMap for large chunks, and a defaultFree
object pointing to the heap that has never been used before.

Gives around 25% performance boost on crypto.js, and bsaically
makes the memory manager invisible in kcachegrind.

Change-Id: I559fb527bcd9e21d4ac265f4d78b8376bfda2522
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>