Flickable: Test case for flicking twice using touches
When you flick twice in rapid succession, in the same direction,
the expected behavior is for flickable to be moving quite fast in the
direction of the flicks.

This test check for a bug where when you flick using touch events
instead of mouse ones, the second flick causes Flickable to immediately halt.

Change-Id: I430515d82499b904a1d2e23402b753873490a2d9
Reviewed-by: Andras Becsi <andras.becsi@digia.com>