Flickable: Fix bug when flicking twice using touches
When you flick twice in rapid succession, in the same direction,
the expected behavior is for flickable to be moving quite fast in the
direction of the flicks.

But when you flicked using touch events instead of mouse ones,
the second flick caused Flickable to immediately halt. This change fixes it.

Change-Id: I02bd02fed30bd4a98959194e04f0a26b4420cd05
Reviewed-by: Andras Becsi <andras.becsi@digia.com>