Fix hang in do-while loops with continue statement in the body
The construct

do {
   ...
   continue;
} while (condition)

Would result in the body basic block ending with a jump to the loop body,
causing an infinite loop.

Instead we have to place the condition into an extra block and set that one as
the target for continue.

This is also covered by ch12/12.14/S12.14_A9_T2, which uses such a loop.

Change-Id: If06de112b338b74b9a49e3d6b51078463645196d
Reviewed-by: Lars Knoll <lars.knoll@digia.com>