Fix parsing of skipped tests
While adding support for running tests that are supposed to fail,
I broke support for skipping tests altogether. After the split()
on the line the result is of course a list, which in the case of
a single "skip" entry needs to be dereferenced instead of appending
a list to a list as a list item.

Change-Id: Ifd27ee4f762ab092a7333b286ca187309cd631ab
Reviewed-by: Lars Knoll <lars.knoll@digia.com>