Exporting required symbols for Qt Quick Designer
These classes are used by qml2puppet.

Qml2Puppet is an external process used by the Qt Quick Designer,
that builds the interface between the designer and Qt Quick/QML
internals.

Qml2Puppet is responsible for emulation and rendering.
The 2 in Qml2Puppet indicates that it is used for Qt Quick 2.0.

This would be nice to have in Qt 5.0.1 to allow testing based
on official packages. Also this helps early adaptors.

QQmlTimer is used to disable timers.

QQuickTextEdit and QQuickTextInput are used to disable the blinking
cursor.

QQuickBehavior is used to deal with behaviors.

QQuickPropertyChanges and QQuickStateGroup are
used to emulate states.

QQuickTransition is used to disable transitions.

Task-number: QTCREATORBUG-8572

Change-Id: I572707d93eda477df945976442efed32ef16df34
Reviewed-by: Kai Koehne <kai.koehne@digia.com>