Fix alignment of Text element with implicit size
In change e1d6b323d474acc5cd70fd9a0b283326316f55a0 there's an
optimization which disables layouts on the QTextDocument, but the
layout has to be re-enabled before we can call idealWidth() since
this will be zero as long as the page size is 0x0. We simply set
the page size to undefined again before calling idealWidth.

Task-number: QTBUG-29262
Change-Id: I6bebe7be4c99fce0de8fc4178b5e1abc81abbe87
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: aavit <eirik.aavitsland@digia.com>