Fix evaluation of lhs in binary expressions
The entire left hand side of a binary expression must be evaluated
before the right hand side. In case of for example

    x !== (x = 1)

it is important to generate the code for looking up "x" from the
lhs before dealing with the rhs, because this overall expression
is supposed to throw a reference error.

Change-Id: I03aee3257ff7b7a60aa789dba9f0445c387a1ad5
Reviewed-by: Lars Knoll <lars.knoll@digia.com>