Dereference the QImage once the corresponding texture is bound.
We do not require a reference here, as the texture is only bound
once. It is the texture factory that is reused (when
QQuickWIndow::setPersistentSceneGraph(false) is set and the window
is hidden and then re-exposed). With this patch it becomes trivial to
create a custom QSGContextPlugin with a texture factory that does
not retain the QImage, freeing up memory. This is useful in the
case of an embedded system with only a single window that
is never hidden or re-exposed.

Change-Id: I1cfa6efc3a2e9e641b456bf90c55d563061757b8
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>