Make sure deleteLater gets run at the "correct" time.
We have a number of places where we delete scene graph objects with
deleteLater() and change the scene graph. The timing of when the
deleteLater happens is very important as we need to complete
the updates to the scene graph first.

In this particular case, a QQuickShaderEffectSource was released,
and an update was scheduled but because animations were running we
were already beginning to render a new frame before the update was
handled, causing a crash.

The only safe place to run deferred deletes is after we have
performed a proper sync with the item scene.

Task-number: QTBUG-29485
Change-Id: I6e93d4e6276fe82d3f4c818781b188e03c46e510
Reviewed-by: Samuel Rødal <samuel.rodal@digia.com>