Fix cmath function usage
Some of the cmath functions in std don't work reliably on all platforms
and on top of that with Visual Studio they are not in the std namespace
at all (but for example prefixed as _isnan). Import MathExtras.h from
WTF that includes all the necessary workarounds, including things like
making pow() with MinGW compliant with what JavaScript needs.

Change-Id: I3d8190ce33919db69b2889b439731a9cf5b2a46e
Reviewed-by: Lars Knoll <lars.knoll@digia.com>