Fix multiple variable assignments
When doing var foo = bar = 42; then we would assign 42 to bar and bar to foo,
resulting in the wrong value for foo if bar was read-only for example.

The spec says in 11.13.1.6 that the rval is to be returned, so we just do
that via the temp we already have.

Change-Id: I44ea895abe4796af10c371baac22c2b26f37b519
Reviewed-by: Lars Knoll <lars.knoll@digia.com>