reference locals in outer functions by index
This makes V4 reference all variables in scopes
that are not the global scope by index. The JIT and
the interpreter walk up the scope chain to get the
correct reference.

Variables are only resolved by name for the global
scope, if the scope contains an eval statement (as
eval can define new variables) and inside with and
catch scopes.

Change-Id: Ib9f9d1a03d50124130aefd169eeb071533ba3520
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>