[masm] Fix function return on arm
Whether we should do ret(n) or ret should depend on whether the caller
provides the pointer to the return value as hidden first parameter or not.
That's the case on ia32 but not on x86-64 or arm, where the first parameter
register is used instead. So the correct preprocessor macro to use here
is ARGUMENTS_IN_REGISTERS instead of VALUE_FITS_IN_REGISTER.

Change-Id: I3a8a8fa316896848baca37626f87ed98c096e14a
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>